-changed PlugVoice.HostTag to be intptr_t to fix crashing issues in larger projects. Thanks FuryGRTX!

+++ Genny/src/base/VSTBase.h
@@ -92,7 +92,7 @@ struct FruityPlugInfo
 };
 
 struct PlugVoice {
-	int HostTag;
+	intptr_t HostTag;
 	PVoiceParams Params;
 	//int Pos[nOsc];
 	int State;
